+++ src/inspect_ai/model/_providers/bedrock.py
@@ -331,6 +331,19 @@ class BedrockAPI(ModelAPI):
     def collapse_assistant_messages(self) -> bool:
         return True
 
+    @override
+    def emulate_reasoning_history(self) -> bool:
+        # claude needs reasoning history emulation because the reasoning signature doesn't
+        # make it all the way through the converse api (so when we try to replay it there is
+        # an error from claude indicating the signature was missing)
+        return self.is_claude()
+
+    def is_gpt_oss(self) -> bool:
+        return "gpt-oss" in self.model_name
+
+    def is_claude(self) -> bool:
+        return "claude" in self.model_name
+
     async def generate(
         self,
         input: list[ChatMessage],
@@ -364,6 +377,14 @@ class BedrockAPI(ModelAPI):
             # Resolve the input messages into converse messages
             system, messages = await converse_messages(input)
 
+            # additional model request fields
+            additionalModelRequestFields: dict[str, Any] = {}
+            if config.top_k:
+                additionalModelRequestFields["top_k"] = config.top_k
+            additionalModelRequestFields = (
+                additionalModelRequestFields | self.reasoning_config(config)
+            )
+
             # Make the request
             request = ConverseClientConverseRequest(
                 modelId=self.model_name,
@@ -375,10 +396,7 @@ class BedrockAPI(ModelAPI):
                     topP=config.top_p,
                     stopSequences=config.stop_seqs,
                 ),
-                additionalModelRequestFields={
-                    "top_k": config.top_k,
-                    **config.model_config,
-                },
+                additionalModelRequestFields=additionalModelRequestFields,
                 toolConfig=tool_config,
             )
 
@@ -418,6 +436,21 @@ class BedrockAPI(ModelAPI):
 
         # return
         return output, model_call(response)
+
+    def reasoning_config(self, config: GenerateConfig) -> dict[str, Any]:
+        if self.is_gpt_oss():
+            if config.reasoning_effort is not None:
+                return {"reasoning_effort": config.reasoning_effort}
+        elif self.is_claude():
+            if config.reasoning_tokens is not None:
+                return {
+                    "reasoning_config": {
+                        "type": "enabled",
+                        "budget_tokens": config.reasoning_tokens,
+                    }
+                }
+
+        return {}
 
 
 async def converse_messages(
@@ -647,6 +680,14 @@ async def converse_contents(
                 )
             elif c.type == "text":
                 result.append(ConverseMessageContent(text=c.text))
+            elif c.type == "reasoning":
+                result.append(
+                    ConverseMessageContent(
+                        reasoningContent=ConverseReasoningContent(
+                            reasoningText=ConverseReasoningText(text=c.reasoning)
+                        )
+                    )
+                )
             else:
                 raise RuntimeError(f"Unsupported content type {c.type}")
         return result